CSP: Move style attribute parsing to the AST
